# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CreateUserPlatformRequest:
    properties:
      alias_name:
        maxLength: 255
        minLength: 1
        type: string
      platform_id:
        type: string
    required:
    - platform_id
    type: object
...
  dto.UpdateUserPlatformRequest:
    properties:
      alias_name:
        maxLength: 255
        minLength: 1
        type: string
      is_active:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive financial management system with smart allocation engine,
    transaction tracking, and detailed reports
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Coin Backend API
  version: "1.0"
paths:
...
  /v1/user-platforms:
    get:
      description: Get all user platforms for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: User platforms retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all user platforms
      tags:
      - User Platforms
    post:
      consumes:
      - application/json
      description: Create a new user platform for the authenticated user
      parameters:
      - description: Platform details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserPlatformRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User platform created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new user platform
      tags:
      - User Platforms
  /v1/user-platforms/{id}:
    delete:
      description: Delete a user platform
      parameters:
      - description: User Platform ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User platform deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete user platform
      tags:
      - User Platforms
    get:
      description: Get a specific user platform by ID
      parameters:
      - description: User Platform ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User platform retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user platform by ID
      tags:
      - User Platforms
    put:
      consumes:
      - application/json
      description: Update a user platform
      parameters:
      - description: User Platform ID
        in: path
        name: id
        required: true
        type: string
      - description: Update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserPlatformRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User platform updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update user platform
      tags:
      - User Platforms
  /v1/user-platforms/dropdown/list:
    get:
      description: Get all user platforms for dropdown with platform data lookup
      produces:
      - application/json
      responses:
        "200":
          description: User platforms retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List user platforms for dropdown
      tags:
      - User Platforms
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"